Add 'Move to Finished Folder' option to GUI
Introduces a new checkbox in the GUI to allow users to enable or disable moving processed source files to a 'finished' folder after processing. Updates configuration handling and processing logic to respect this user setting in both single and batch modes. Also adds corresponding help text for the new option.